+++ src/hooks/MenuLayer.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Geode/modify/MenuLayer.hpp>
+
+class $modify(HookedMenuLayer, MenuLayer) {
+    static cocos2d::CCScene* scene(bool isVideoOptionsOpen);
+};